added LDRH/LDRSH/LDRSB immediate/register post-indexed addressing mode instructions
Former-commit-id: 4ed17acb6f5916fc2c5b59e5290449be1174c6d2

+++ src/jumptable/jumptable-arm-config.cpp
@@ -230,7 +230,7 @@ uint32_t addressing_mode_3_immediate_offset(ARM7TDMI* cpu, uint32_t opcode) {
 
 @LOCAL()
 uint32_t addressing_mode_3_register_offset(ARM7TDMI* cpu, uint32_t opcode) {
-    uint8_t offset = cpu->regs[get_nth_bits(opcode, 0, 4)];
+    uint32_t offset = cpu->regs[get_nth_bits(opcode, 0, 4)];
     if (get_nth_bit(opcode, 23)) return cpu->regs[get_nth_bits(opcode, 16, 20)] + offset;
     else                         return cpu->regs[get_nth_bits(opcode, 16, 20)] - offset;
 }
@@ -246,9 +246,9 @@ uint32_t addressing_mode_3_immediate_preindexed(ARM7TDMI* cpu, uint32_t opcode) 
 
 @LOCAL()
 uint32_t addressing_mode_3_register_preindexed(ARM7TDMI* cpu, uint32_t opcode) {
-    uint8_t offset = cpu->regs[get_nth_bits(opcode, 0, 4)];
-    if (get_nth_bit(opcode, 23)) cpu->regs[get_nth_bits(opcode, 16, 20)] += (int32_t) offset;
-    else                         cpu->regs[get_nth_bits(opcode, 16, 20)] -= (int32_t) offset;
+    uint32_t offset = cpu->regs[get_nth_bits(opcode, 0, 4)];
+    if (get_nth_bit(opcode, 23)) cpu->regs[get_nth_bits(opcode, 16, 20)] += offset;
+    else                         cpu->regs[get_nth_bits(opcode, 16, 20)] -= offset;
     return cpu->regs[get_nth_bits(opcode, 16, 20)];
 }
 
@@ -264,10 +264,10 @@ uint32_t addressing_mode_3_immediate_postindexed(ARM7TDMI* cpu, uint32_t opcode)
 
 @LOCAL()
 uint32_t addressing_mode_3_register_postindexed(ARM7TDMI* cpu, uint32_t opcode) {
-    uint8_t  offset  = cpu->regs[get_nth_bits(opcode, 0, 4)];
+    uint32_t offset  = cpu->regs[get_nth_bits(opcode, 0, 4)];
     uint32_t address = cpu->regs[get_nth_bits(opcode, 16, 20)];
-    if (get_nth_bit(opcode, 23)) cpu->regs[get_nth_bits(opcode, 16, 20)] += (int32_t) offset;
-    else                         cpu->regs[get_nth_bits(opcode, 16, 20)] -= (int32_t) offset;
+    if (get_nth_bit(opcode, 23)) cpu->regs[get_nth_bits(opcode, 16, 20)] += offset;
+    else                         cpu->regs[get_nth_bits(opcode, 16, 20)] -= offset;
     return address;
 }
